sensor code updated

+++ src/main/java/edu/ius/robotics/robots/IRobot.java
@@ -2,5 +2,5 @@ package edu.ius.robotics.robots;
 
 public interface IRobot
 {
-	public void sensorEvent(int[] sensorData);
+	public void sensorEvent(byte[] sensorData);
 }
